+++ src/senaite/core/exportimport/setupdata/__init__.py
@@ -660,11 +660,12 @@ class Client_Contacts(WorksheetImporter):
                             "fullname": fullname,
                         }
                     )
+                    # This will add the user to a client specific group which
+                    # has the "Client" role assigned
+                    contact.setUser(username)
+
                 except Exception as msg:
                     logger.info("Error adding user (%s): %s" % (msg, username))
-                contact.aq_parent.manage_setLocalRoles(
-                    row['Username'], ['Owner', ])
-                contact.reindexObject()
 
 
 class Container_Types(WorksheetImporter):
@@ -2112,7 +2113,9 @@ class Setup(WorksheetImporter):
     def to_string_vocab_value(self, field, value):
         vocabulary = field.vocabulary
         if type(vocabulary) is str:
-            vocabulary = getFromString(api.get_setup(), vocabulary)
+            # Use bika_setup for vocabulary access (has subfolders)
+            bika_setup = api.get_portal().get("bika_setup")
+            vocabulary = getFromString(bika_setup, vocabulary)
         else:
             vocabulary = vocabulary.items()